Problem: gaslighter doesn't have subcommands for its various activities
Solution: add the stubs subcommands for `service`,
`blockchain regression`, `cli`, `ffi`, `crate`. Also refactor the
original gaslighter to fulfill the `ffi` functionality.
